+++ src/hooks/useLandManagerRegionData.ts
@@ -1,11 +1,13 @@
 "use client";
 
 import {
-  getRegionStakedDEC,
   getRentedCardsList,
-  RegionDECInfo,
   RentedCardsList,
 } from "@/lib/backend/actions/land-manager/rental-actions";
+import {
+  RegionDECInfo,
+  getRegionStakedDEC,
+} from "@/lib/backend/actions/land-manager/dec-power-actions";
 import { getWorkerEligibility } from "@/lib/backend/actions/land-manager/worker-actions";
 import { WorkerEligibilityResult } from "@/types/landManager";
 import { useEffect, useState } from "react";